fixed fbsp vertex lighting the lame way. still needs a glsl method. fixed shader lighting levels due to overbright methods. fixed scaling. misc other fixes and tweaks.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4469 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/cl_screen.c
@@ -914,7 +914,7 @@ void CL_Sbar_Callback(struct cvar_s *var, char *oldvalue)
 {
 }
 
-void SCR_CrosshairPosition(playerview_t *pview, int *x, int *y)
+void SCR_CrosshairPosition(playerview_t *pview, float *x, float *y)
 {
 	extern cvar_t cl_crossx, cl_crossy, crosshaircorrect, v_viewheight;
